Remove cancellation from inertia calculation to improve accuracy (#955)
I was computing the shape inertia about the body origin and then
shifting to the center of mass. But this involves quadratic math with
bad cancellation effects.

I'm now buffering the shape inertias about their centroids and then
shifting them to the final body center of mass. This removes the
quadratic cancellation.

Fixes #949